fix: try fixing issues based on feedback, plus split electron.js to avoid large file
